Some more fixes for Linux
Add global executable overrides as needed then starting x64 tasks
referencing a flatpak installed executable.

+++ src/main/java/org/cbm/ant/cbm/disk/CBMDiskUtil.java
@@ -243,7 +243,7 @@ public class CBMDiskUtil
     {
         for (int i = 0; i < PETSCII_MAPPING.length; i += 1)
         {
-            ASCII_MAPPING.put(PETSCII_MAPPING[i], Byte.valueOf((byte) (i & 0xff)));
+            ASCII_MAPPING.put(PETSCII_MAPPING[i], (byte) (i & 0xff));
         }
     }
